backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
pydantic>=2.10.0
python-dotenv>=1.0.1
python-multipart>=0.0.18
openai-whisper>=20240930
imageio-ffmpeg>=0.5.1
